Introduce a new package `@adeira/sx-design-headless`
This will allow us to share the SX Design core across multiple design systems.
