+++ scripts/publishedPackages.json
@@ -16,6 +16,7 @@
   "@adeira/relay",
   "@adeira/sx",
   "@adeira/sx-design",
+  "@adeira/sx-design-headless",
   "@adeira/sx-design-nextjs",
   "@adeira/sx-jest-snapshot-serializer",
   "eslint-plugin-adeira",
