+++ src/sx-design-headless/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as Button } from './src/Button/Button';
+export type { Props } from './src/Button/Button';
